Lots of progress! It's mostly working!
Fixed the crash: a nil texture ref.
Added some translations that are not yet in use.
Added some keywords to items.
Hacked up one of the scripts from "Campfire Animations"
to make that mod work with this one, if you have both Campfire
and the animations mod installed.

It's starting to feel pretty cool! Or hot. You know.