Rework for handling Campfire integration
Add states back to the papyrus script so we can remove all
the logic from the perk activation script. The states now
manage what we do on activation and DAK activation.

There are too many properties on the activators now, but
I can clean those up once this settles down. There's also
a problem with the DAK activation text: it's still "Cook"
which is now wrong, but there is no single word that will
work for all of the alt activations. I might need different
perks.